Implemented the basics of running the game environment, console player, etc. The game adds ConsolePlayers, asks for mulligans, can play lands, and can somehwat buggily play creatures.
Next on the todo list is to finish casting creatures, then add creature combat.
I need to consider how granular I want the actions to be - should tapping lands be part of casting a spell?  Do I want to ask what colors to use for casting the spell?
The code is really ugly right now.  Probably not the right time to be worrying about it.